==> code/01_initialize-dirs.R <==
library(fs)
library(glue)


dir_create("data/input")
dir_create("data/input/surveys")
dir_create("data/input/surveys/538")
dir_create("data/input/elections")
dir_create("stan-src")
dir_create("figures")
dir_create("numbers")
dir_create("tables")


dir_create("data/output")
dir_create("data/output/538-polls_cleaned")
dir_create("data/output/HDSR-models")
